:recycle: [refactor] 스테이지 캐치 프로그래스바 위젯화
